* iseq.c (ruby_iseq_disasm_insn): suppress warnings on platforms which   int size differs from pointer size.
* ext/openssl/ossl_asn1.c (ossl_asn1_get_asn1type): ditto

* ext/syck/rubyext.c (rb_syck_err_handler),
  (syck_default_error_handler): ditto.


git-svn-id: svn+ssh://ci.ruby-lang.org/ruby/trunk@18283 b2dd03c8-39d4-4d8f-98ff-823fe69b080e

+++ iseq.c
@@ -733,7 +733,7 @@ ruby_iseq_disasm_insn(VALUE ret, VALUE *iseq, int pos,
     }
     else {
 	rb_str_catf(str, "%04d %-16.*s ", pos,
-		    strcspn(insn_name_buff, "_"), insn_name_buff);
+		    (int)strcspn(insn_name_buff, "_"), insn_name_buff);
     }
 
     for (j = 0; types[j]; j++) {
